backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
Pillow
aiofiles
pytest
httpx